% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annData.R
\name{annData}
\alias{annData}
\title{annData}
\usage{
annData(
  path,
  notes,
  date = Sys.Date(),
  recursive = TRUE,
  md5 = FALSE,
  skip = "*.md|meta.yml",
  force = FALSE,
  ...
)
}
\arguments{
\item{path}{The data path to annotate.}

\item{notes}{User assigned notes/keywords to annotate the data and
be used for keywords matching in \code{dataSearch(keywords = )}.}

\item{date}{The date of the data.}

\item{recursive}{Whether to annotate all data recursively.}

\item{md5}{Whether to generate md5 values for all files.}

\item{skip}{Patter to skip files in the path.}

\item{force}{Whether to force regenerate meta.yml.}

\item{...}{The other options from \code{list.files}}
}
\description{
Add annotation or meta information to existing data
}
