% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSearch.R
\name{dataSearch}
\alias{dataSearch}
\title{dataSearch
search data in local data caching system}
\usage{
dataSearch(keywords = character(), cachePath = "ReUseData", RUDcloud = TRUE)
}
\arguments{
\item{keywords}{character vector of keywords to be matched to the
local datasets. It matches the "notes" when generating the data
using \code{getData(notes = )}. Keywords can be a tag with the data
in \verb{#tag} format. If not specified, function returns the full
data list.}

\item{cachePath}{A character string for the data cache. Must
match the one specified in \code{dataUpdate()}. Default is
"ReUseData".}
}
\value{
a \code{dataHub} object containing the information about local
data cache, e.g., data name, data path, etc.
}
\description{
dataSearch
search data in local data caching system
}
\examples{
dataSearch()
dataSearch(c("gencode")) 
dataSearch("#gatk")

}
