% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{dataTag}
\alias{dataTag}
\alias{dataTag,dataHub-method}
\alias{dataTag<-}
\alias{dataTag<-,dataHub-method}
\title{dataTagb}
\usage{
dataTag(object)

\S4method{dataTag}{dataHub}(object)

dataTag(object, append = TRUE) <- value

\S4method{dataTag}{dataHub}(object, append = FALSE) <- value
}
\arguments{
\item{object}{A \code{dataHub} object.}

\item{append}{Whether to append new tag or replace all tags.}

\item{value}{The value(s) to assign to the \code{dataHub} object.}
}
\description{
Show tag column of a \code{dataHub} object.
}
