% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{dataTags}
\alias{dataTags}
\alias{dataTags,dataHub-method}
\alias{dataTags<-}
\alias{dataTags<-,dataHub-method}
\title{show tags}
\usage{
dataTags(object)

\S4method{dataTags}{dataHub}(object)

dataTags(object) <- value

\S4method{dataTags}{dataHub}(object) <- value
}
\arguments{
\item{object}{A \code{dataHub} object.}

\item{value}{A data.frame of fpath and tags to update the metadata.}
}
\description{
Show a data.frame of file paths and tags.
}
