% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUpdate.R
\name{dataUpdate}
\alias{dataUpdate}
\title{update the local data records by reading the yaml files in the
specified directory recursively.}
\usage{
dataUpdate(dir, cachePath = "ReUseData", outMeta = FALSE, keepTags = TRUE)
}
\arguments{
\item{dir}{a character string for the directory where all data are
saved. Data information will be collected recursively within
this directory.}

\item{cachePath}{the cache path for recording all available
datasets. Default is "ReUseData".}
}
\value{
a \code{dataHub} object containing the information about local
data cache, e.g., data name, data path, etc.
}
\description{
update the local data records by reading the yaml files in the
specified directory recursively.
}
\examples{
## dataUpdate(dir = "~/workspace/SharedData")
}
