% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_data}
\alias{meta_data}
\title{meta_data()}
\usage{
meta_data(dir = "", outdir = dir)
}
\arguments{
\item{dir}{The path to the data folder.}

\item{outdir}{The path the the output csv file.}
}
\value{
a csv file with columns: yml file name, Output path, Notes, Version, Date, etc.
}
\description{
Functions to generate the meta csv file for local cached dataset.
}
\examples{
meta_data("../res")
}
