% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{recipeHub}
\alias{recipeHub}
\alias{show,recipeHub-method}
\title{recipeHub}
\usage{
recipeHub(BFC)

recipeHub(BFC)

\S4method{show}{recipeHub}(object)
}
\arguments{
\item{BFC}{A BiocFileCache object created for recipe and recipes.}

\item{x}{A \code{recipeHub} object}
}
\value{
recipeHub: a \code{recipeHub} object.
}
\description{
\code{recipeHub} class, constructor, and methods.
}
