% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipeLoad.R
\name{recipeLoad}
\alias{recipeLoad}
\title{recipeLoad}
\usage{
recipeLoad(rname, cachePath = "ReUseDataRecipe", env = .GlobalEnv)
}
\arguments{
\item{rname}{The name or filepath of data recipe (\code{rname} or
\code{fpath} column from the \code{bfc} object returned from
\code{recipeSearch}). It can also be an url or a github repo.}

\item{cachePath}{The cache path of the BiocFileCache object to
store the ReUseData recipes.}

\item{env}{The R enviroment to export to. The default is
\code{.GlobalEnv}.}
}
\value{
A \code{cwlProcess} object. For advanced/composite data recipes,
the basic data recipes involved will also be loaded.
}
\description{
To load a data recipe into R environment.
}
\examples{
\dontrun{
rcp <- recipeSearch("gencode")
rcp$rname
rcp$rpath
gtf <- recipeLoad("gencode_gtf")
gtf <- recipeLoad(rcp$rpath[rcp$rname == "gencode_gtf"])  ## equivalent
gtf
}
}
