% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipeSearch.R
\name{recipeSearch}
\alias{recipeSearch}
\title{search recipes on GitHub repository}
\usage{
recipeSearch(keywords, cachePath = "ReUseDataRecipe")
}
\arguments{
\item{keywords}{keywords}

\item{cachePath}{ReUseData by default}
}
\value{
return a DataHub object to be defined?
}
\description{
search recipes on GitHub repository
}
\examples{
recipeSearch("gencode")
}
