% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReUseData-package.R
\docType{package}
\name{ReUseData}
\alias{ReUseData}
\title{ReUseData}
\description{
ReUseData is an \emph{R/Bioconductor} software tool to provide a
systematic and versatile approach for standardized and reproducible
data management. ReUseData facilitates transformation of shell or
other ad hoc scripts for data preprocessing into workflow-based
data recipes. Evaluation of data recipes generate curated data
files in their generic formats (e.g., VCF, bed). Both recipes and
data are cached using database infrastructure for easy data
management and reuse. Prebuilt data recipes are available through
ReUseData portal ("https://rcwl.org/dataRecipes/") with full
annotation and user instructions. Pregenerated data are available
through ReUseData cloud bucket that is directly downloadable
through "getCloudData()".
}
