% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCWLData.R
\name{runCWLData}
\alias{runCWLData}
\title{Run Rcwl for data recipes}
\usage{
runCWLData(cwl, outdir, prefix, notes = Sys.Date(), ...)
}
\arguments{
\item{cwl}{An Rcwl object with value assigned.}

\item{prefix}{The prefix for cwl and yml files.}

\item{notes}{The notes or annotation to keep for the data.}

\item{...}{Other options from \code{runCWL}.}
}
\value{
A list including yml meta file.
}
\description{
Run Rcwl for data recipes
}
