% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\alias{tag,dataHub-method}
\alias{tag<-}
\alias{tag<-,dataHub-method}
\title{tag}
\usage{
tag(object)

\S4method{tag}{dataHub}(object)

tag(object, append = TRUE) <- value

\S4method{tag}{dataHub}(object, append = FALSE) <- value
}
\arguments{
\item{object}{A \code{dataHub} object.}

\item{append}{Whether to append new tag or replace all tags.}

\item{value}{The value(s) to assign to the \code{dataHub} object.}
}
\description{
Show tag column of a \code{dataHub} object.
}
