% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tags}
\alias{tags}
\alias{tags,dataHub-method}
\alias{tags<-}
\alias{tags<-,dataHub-method}
\title{show tags}
\usage{
tags(object)

\S4method{tags}{dataHub}(object)

tags(object) <- value

\S4method{tags}{dataHub}(object) <- value
}
\arguments{
\item{object}{A \code{dataHub} object.}

\item{value}{A data.frame of fpath and tags to update the metadata.}
}
\description{
Show a data.frame of file paths and tags.
}
