% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{dataHub}
\alias{dataHub}
\alias{show,dataHub-method}
\alias{dataParams}
\alias{dataNotes}
\title{dataHub}
\usage{
dataHub(BFC)

dataHub(BFC)

\S4method{show}{dataHub}(object)

dataParams(object)

dataNotes(object)
}
\arguments{
\item{BFC}{A BiocFileCache object created for data and recipes.}

\item{x}{A \code{dataHub} object}
}
\value{
dataHub: a \code{dataHub} object.

dataParams: the "parameter" values for the \code{dataHub} object.

dataNotes: the "notes" for the \code{dataHub} object.
}
\description{
\code{dataHub} class, constructor, and methods.
}
