% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHub-class.R
\name{dataHub}
\alias{dataHub}
\alias{show,dataHub-method}
\alias{dataNames}
\alias{dataParams}
\alias{dataNotes}
\alias{dataPaths}
\alias{dataYml}
\alias{dataTags}
\alias{dataTags,dataHub-method}
\alias{dataTags<-}
\alias{dataTags<-,dataHub-method}
\alias{[,dataHub,ANY,ANY,ANY-method}
\alias{c,dataHub-method}
\alias{toList}
\title{dataHub Class}
\usage{
dataHub(BFC)

dataHub(BFC)

\S4method{show}{dataHub}(object)

dataNames(object)

dataParams(object)

dataNotes(object)

dataPaths(object)

dataYml(object)

dataTags(object)

\S4method{dataTags}{dataHub}(object)

dataTags(object, append = TRUE) <- value

\S4method{dataTags}{dataHub}(object, append = FALSE) <- value

\S4method{[}{dataHub,ANY,ANY,ANY}(x, i)

\S4method{c}{dataHub}(x, ...)

toList(x, format = c("list", "json", "yaml"), type = NULL)
}
\arguments{
\item{BFC}{A BiocFileCache object created for data and recipes.}

\item{object}{A \code{dataHub} object.}

\item{append}{Whether to append new tag or replace all tags.}

\item{value}{The tag values to assign to datasets in \code{dataHub} object.}

\item{x}{A \code{dataHub} object.}

\item{i}{The integer index of the \code{dataHub} object, or a logical
vector same length as the \code{dataHub} object.}

\item{...}{More \code{dataHub} objects to combine.}

\item{format}{can be "list", "json" or "yaml". Supports partial
match. Default is list.}

\item{type}{The type of workflow input list, such as cwl.}
}
\value{
dataHub: a \code{dataHub} object.

dataNames: the names of datasets in \code{dataHub} object.

dataParams: the data recipe parameter values for datasets in \code{dataHub} object.

dataNotes: the notes of datasets in \code{dataHub} object.

dataPaths: the file paths of datasets in \code{dataHub} object.

dataYml: the yaml file paths of datasets in \code{dataHub} object.

dataTags: the tags of datasets in \code{dataHub} object.

toList: A list of datasets in specific format.
}
\description{
\code{dataHub} class, constructor, and methods.
}
\examples{
outdir <- file.path(tempdir(), "SharedData")
dataUpdate(outdir, cloud = TRUE)
dd <- dataSearch(c("liftover", "GRCh38"))
dataNames(dd)
dataParams(dd)
dataNotes(dd)
dataTags(dd)
dataYml(dd)
toList(dd)
toList(dd, format = "json")
toList(dd, format = "yaml")
}
