% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSearch.R
\name{dataSearch}
\alias{dataSearch}
\title{search recipes on GitHub repository}
\usage{
dataSearch(keywords = character(), cachePath = "ReUseData")
}
\arguments{
\item{keywords}{character vector of keywords to be matched to the
local datasets. It matches the "notes" when generating the data
using \code{getData(notes = )}. Keywords can be a tag with the data
in \verb{#tag} format. If not specified, function returns the full
data list.}

\item{cachePath}{ReUseData by default}
}
\value{
a \code{dataHub} object containing the information about local
data cache, e.g., data name, data path, etc.
}
\description{
search recipes on GitHub repository
}
\examples{
## dataSearch()
## dataSearch(c("gencode", "human", "38"))
## dataSearch("#samtools")
}
