% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSearch.R
\name{dataSearch}
\alias{dataSearch}
\title{search recipes on GitHub repository}
\usage{
dataSearch(keywords, cachePath = "ReUseData")
}
\arguments{
\item{keywords}{keywords}

\item{cachePath}{ReUseData by default}
}
\value{
a \code{dataHub} object containing the information about local
data cache, e.g., data name, data path, etc.
}
\description{
search recipes on GitHub repository
}
\examples{
dataSearch(c("gencode", "human", "38"))
}
