% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUpdate.R
\name{dataUpdate}
\alias{dataUpdate}
\title{update the data records by reading the .yml files using BiocFileCache}
\usage{
dataUpdate(dir, cachePath = "ReUseData", outMeta = FALSE, keepTags = TRUE)
}
\arguments{
\item{dir}{a character string for the path to save the data.}

\item{cachePath}{the cache path for recording all available
datasets. Default is "ReUseData".}
}
\value{
a \code{dataHub} object containing the information about local
data cache, e.g., data name, data path, etc.
}
\description{
update the data records by reading the .yml files using BiocFileCache
}
\examples{
## dataUpdate(dir = "~/workspace/SharedData")
}
