% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData: Generate the curated dataset from an provided data recipe.}
\usage{
getData(cwl, outdir, prefix, version = "", notes = "", docker = TRUE, ...)
}
\arguments{
\item{cwl}{the \code{cwlProcess} object from \code{searchData()} or
\code{loadData()}.}

\item{outdir}{The directory to store the outputs. Will
automatically create if not exist.}

\item{prefix}{The string that labels the data which will be added
as prefix to the 4 output files. e.g., "gencode_human_38".}

\item{version}{User assigned version for the data generated. e.g.,
"38", "v1.0", etc.}

\item{notes}{User assigned notes/keywords to retrieve the data by
\code{searchData()} function.  e.g., "human", "reference genome",
"gencode", "gtf", etc.}

\item{docker}{Default is TRUE.}

\item{...}{Arguments to be passed into \code{Rcwl:runCWL()}.}
}
\value{
The data files and 4 meta files: \code{.cwl}: The cwl script
that was internally run to download the data; \code{.yml}: the
values for the \code{cwlProcess} object, and some user added notes,
versions etc. \code{.sh}: The script for data downloading and
curation.  \code{.md}: ... (to-be-filled)
}
\description{
getData: Generate the curated dataset from an provided data recipe.
}
