% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(
  cwl,
  outdir,
  prefix,
  version = character(),
  notes = c(),
  docker = TRUE,
  ...
)
}
\arguments{
\item{cwl}{the data recipe in \code{cwlProcess} S4 class.}

\item{outdir}{Character string specifying the directory to store
the output files. Will automatically create if not exist or
provided.}

\item{prefix}{Character string that labels the data which will be
added as prefix to the automatically generated annotation and
workflow files. e.g., "gencode_human_38".}

\item{version}{Character string of user specified version for the
data generated for local data management. Default is
empty. When provided, it will be pasted with \code{prefix} argument
using "_" to serve as the final prefix for output files. E.g.,
"gencode_human_38_v1.0".}

\item{notes}{User assigned notes/keywords to annotate the data and
be used for keywords matching in \code{dataSearch(keywords = )}.}

\item{docker}{Whether to use docker when evaluating the data recipe
as a CWL workflow. Default is TRUE.}

\item{...}{Arguments to be passed into \code{Rcwl:runCWL()}.}
}
\value{
The data files and 4 meta files: \code{.cwl}: The cwl script
that was internally run to get the data; \code{.yml}: the input
parameter values for the data recipe and user specified data
annoation notes, versions etc; \code{.sh}: The script for data
processing; \code{.md}: checksum file to verify the integrity of
generated data files.
}
\description{
Evaluation of data recipes to generate curated dataset of interest.
}
\examples{
rcp <- recipeLoad("gencode_transcripts")
rcp$species <- "human"
rcp$version <- "42"
outdir <- file.path(tempdir(), "SharedData")
res <- getData(rcp,
        outdir = outdir, 
        prefix = "gencode_annotation_human_42",
        notes = c("gencode", "human", "42"),
        showLog = TRUE)
dir(outdir)

}
