% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_data}
\alias{meta_data}
\title{meta_data}
\usage{
meta_data(dir = "")
}
\arguments{
\item{dir}{The path to the shared data folder.}
}
\value{
a \code{data.frame} with yml file name, parameter values, data
file paths, date, and user-specified notes, version when
generating the data with \code{getData()}.
}
\description{
Functions to generate the meta csv file for local cached dataset.
}
\examples{
## meta_data("../SharedData")
}
