% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipeHub-class.R
\name{recipeHub}
\alias{recipeHub}
\alias{show,recipeHub-method}
\alias{[,recipeHub,ANY,ANY,ANY-method}
\alias{recipeNames}
\title{recipeHub}
\usage{
recipeHub(BFC)

recipeHub(BFC)

\S4method{show}{recipeHub}(object)

\S4method{[}{recipeHub,ANY,ANY,ANY}(x, i)

recipeNames(object)
}
\arguments{
\item{BFC}{A BiocFileCache object created for recipe and recipes.}

\item{object}{The \code{recipeHub} object}

\item{x}{The \code{recipeHub} object}

\item{i}{The integer index of the \code{recipeHub} object}
}
\value{
recipeHub: a \code{recipeHub} object.

[: A \code{recipeHub} object that was subsetted.

recipeNames: the recipe names for the \code{recipeHub} object.
}
\description{
\code{recipeHub} class, constructor, and methods.
}
\examples{
rcps <- recipeSearch(c("gencode"))
## rcp1 <- rcps[1]
## recipeNames(rcp1)
}
