% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipeLoad.R
\name{recipeLoad}
\alias{recipeLoad}
\title{recipeLoad}
\usage{
recipeLoad(
  rcp = c(),
  cachePath = "ReUseDataRecipe",
  env = .GlobalEnv,
  return = FALSE
)
}
\arguments{
\item{rcp}{The (vector of) character string of recipe name or
filepath (\code{recipeNames()} or \code{mcols()$fpath} column of the
\code{recipeHub} object returned from \code{recipeSearch}).}

\item{cachePath}{A character string for the recipe cache. Must
match the one specified in \code{recipeUpdate()}. Default is
"ReUseDataRecipe".}

\item{env}{The R enviroment to export to. Default is \code{.GlobalEnv}.}

\item{return}{Whether to use the original recipe name. Default is
FALSE, where user need to assign a variable name to the
recipe. e.g., \code{rcp1 <- recipeLoad()}. If TRUE, it loads the
recipe and uses its original name, and user doesn't need to
assign a new name. e.g., \code{recipeLoad(return=TRUE)}. If multiple
recipes are to be loaded, \code{return=TRUE} must be used.}
}
\value{
A data recipe of \code{cwlProcess} S4 class, which is ready to
be evaluated in \emph{R}.
}
\description{
To load data recipe(s) into R environment.
}
\examples{
########################
## Load single recipe
########################

recipeSearch("liftover")
rcp <- recipeLoad("ensembl_liftover")
inputs(rcp)
rm(rcp)

recipeLoad("gencode_annotation", return=TRUE)
inputs(gencode_annotation)
rm(gencode_annotation)

#########################
## Load multiple recipes
#########################

rcphub <- recipeSearch("gencode")
recipeNames(rcphub)
recipeLoad(recipeNames(rcphub), return=TRUE)
inputs(gencode_transcripts)
}
