% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipeSearch.R
\name{recipeSearch}
\alias{recipeSearch}
\title{recipeSearch}
\usage{
recipeSearch(keywords = character(), cachePath = "ReUseDataRecipe")
}
\arguments{
\item{keywords}{character vector of keywords to be matched to the
recipe names. If not specified, function returns the full
recipe list.}

\item{cachePath}{A character string for the recipe cache. Must
match the one specified in \code{recipeUpdate()}. Default is
"ReUseDataRecipe".}
}
\value{
A \code{recipeHub} object.
}
\description{
Search existing data recipes.
}
\examples{
recipeSearch()
recipeSearch("gencode")
recipeSearch(c("STAR", "index"))
}
