% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipeUpdate.R
\name{recipeUpdate}
\alias{recipeUpdate}
\title{recipeUpdate}
\usage{
recipeUpdate(
  cachePath = "ReUseDataRecipe",
  force = FALSE,
  repos = "rworkflow/ReUseDataRecipe"
)
}
\arguments{
\item{cachePath}{The cache path of the BiocFileCache object to
store the ReUseData recipes. "ReUseDataRecipe" by default.}

\item{force}{Whether to clean existing recipes cache. Default is
FALSE.}

\item{repos}{The GitHub repository for data recipes. By default, it
reads the "rworkflows/ReUseDataRecipes" GitHub
repository. Users can also specify a GitHub repo for their
private data recipes.}
}
\description{
Function to sync and get the most updated data recipes through the
pubic "rworkflows/ReUseDataRecipes" GitHub repository or user
specified private GitHub repo.
}
\examples{
\dontrun{
rcps <- recipeUpdate()
rcps
}
}
