% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipeUpdate.R
\name{recipeUpdate}
\alias{recipeUpdate}
\title{recipeUpdate}
\usage{
recipeUpdate(cachePath = "ReUseDataRecipe", force = FALSE)
}
\arguments{
\item{cachePath}{The cache path of the BiocFileCache object to
store the ReUseData recipes.}

\item{force}{Whether to clean existing recipes cache.}

\item{branch}{The branch of github recipes repository. It can be
"master" and "dev". "force = TRUE" is recommended when swithing
branch.}
}
\description{
Function to sync and get the most updated data recipes from the
"rworkflows/ReUseDataRecipes" GitHub repository.
}
\examples{
\dontrun{
tools <- UpdateRecipe()
}
}
